requirements.txt

pillow>=10.0.0

requests>=2.31.0

qdrant-client>=1.12.0,<1.13.0

sentence-transformers>=2.2.0
sentencepiece>=0.1.99

numpy>=1.24.0

tqdm>=4.66.0

torch>=2.0.0
transformers>=4.35.0

fugashi>=1.5.0
unidic-lite>=1.0.8

ffmpeg-python>=0.2.0

mutagen>=1.47.0
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.1
openai-whisper>=20231117

panns-inference>=0.1.1

crepe>=0.0.14

# SVG rasterization for image logo analysis
cairosvg>=2.7.0

tensorflow-macos==2.16.2; platform_system == "Darwin" and platform_machine == "arm64"
tensorflow-cpu>=2.12.0; platform_system != "Darwin" or platform_machine != "arm64"

tf-keras>=2.16.0
